requirements.txt
# 基于深度学习的图像风格迁移系统依赖

# Web框架
flask==2.0.1
werkzeug==2.0.1

# 数据处理
numpy==1.21.0
pillow==8.3.1

# 深度学习框架
torch==1.9.0
torchvision==0.10.0

# 其他工具
uuid==1.30